remove mr. anderson support from encounter
