+++ data-query-tests/src/test/java/gov/va/api/health/dataquery/tests/dstu2/DataQueryValidateIT.java
@@ -15,7 +15,6 @@ import gov.va.api.health.dataquery.tests.SystemDefinitions;
 import gov.va.api.health.dataquery.tests.TestClients;
 import gov.va.api.health.dataquery.tests.TestIds;
 import gov.va.api.health.dstu2.api.bundle.AbstractBundle;
-import gov.va.api.health.dstu2.api.resources.Encounter;
 import gov.va.api.health.dstu2.api.resources.Location;
 import gov.va.api.health.dstu2.api.resources.MedicationDispense;
 import gov.va.api.health.dstu2.api.resources.OperationOutcome;
@@ -89,12 +88,17 @@ public class DataQueryValidateIT {
     validate("DiagnosticReport", ids.diagnosticReport(), DiagnosticReport.Bundle.class);
   }
 
-  @Test
-  @Category(Local.class)
-  public void validateEncounter() {
-    validate("Encounter", ids.encounter(), Encounter.Bundle.class);
-  }
-
+  /*
+    ----------------------------------------------------------------------------------
+    Disabling Validate tests until the resource becomes available in datamart
+    and the controller/transformer are implemented.
+    ----------------------------------------------------------------------------------
+    @Test
+    @Category(Local.class)
+    public void validateEncounter() {
+      validate("Encounter", ids.encounter(), Encounter.Bundle.class);
+    }
+  */
   @Test
   @Category(Local.class)
   public void validateImmunization() {
