+++ data-query-tests/src/test/java/gov/va/api/health/dataquery/tests/dstu2/EncounterIT.java
@@ -1,7 +1,6 @@
 package gov.va.api.health.dataquery.tests.dstu2;
 
 import gov.va.api.health.dataquery.tests.ResourceVerifier;
-import gov.va.api.health.dstu2.api.resources.Encounter;
 import gov.va.api.health.dstu2.api.resources.OperationOutcome;
 import gov.va.api.health.sentinel.categories.Local;
 import lombok.experimental.Delegate;
@@ -17,9 +16,9 @@ public class EncounterIT {
   @Test
   public void advanced() {
     verifier.verifyAll(
-        test(200, Encounter.Bundle.class, "Encounter?_id={id}", verifier.ids().encounter()),
-        test(404, OperationOutcome.class, "Encounter?_id={id}", verifier.ids().unknown()),
-        test(200, Encounter.Bundle.class, "Encounter?identifier={id}", verifier.ids().encounter()));
+        test(501, OperationOutcome.class, "Encounter?_id={id}", verifier.ids().encounter()),
+        test(501, OperationOutcome.class, "Encounter?_id={id}", verifier.ids().unknown()),
+        test(501, OperationOutcome.class, "Encounter?identifier={id}", verifier.ids().encounter()));
   }
 
   @Category({Local.class
@@ -28,7 +27,7 @@ public class EncounterIT {
   @Test
   public void basic() {
     verifier.verifyAll(
-        test(200, Encounter.class, "Encounter/{id}", verifier.ids().encounter()),
-        test(404, OperationOutcome.class, "Encounter/{id}", verifier.ids().unknown()));
+        test(501, OperationOutcome.class, "Encounter/{id}", verifier.ids().encounter()),
+        test(501, OperationOutcome.class, "Encounter/{id}", verifier.ids().unknown()));
   }
 }
