--- data-query/src/main/java/gov/va/api/health/dataquery/service/controller/encounter/EncounterController.java
@@ -1,128 +1,0 @@
-package gov.va.api.health.dataquery.service.controller.encounter;
-
-import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.firstPayloadItem;
-import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.hasPayload;
-
-import gov.va.api.health.dataquery.service.controller.CountParameter;
-import gov.va.api.health.dataquery.service.controller.Dstu2Bundler;
-import gov.va.api.health.dataquery.service.controller.Dstu2Bundler.BundleContext;
-import gov.va.api.health.dataquery.service.controller.Dstu2Validator;
-import gov.va.api.health.dataquery.service.controller.PageLinks.LinkConfig;
-import gov.va.api.health.dataquery.service.controller.Parameters;
-import gov.va.api.health.dataquery.service.mranderson.client.MrAndersonClient;
-import gov.va.api.health.dataquery.service.mranderson.client.Query;
-import gov.va.api.health.dstu2.api.resources.Encounter;
-import gov.va.api.health.dstu2.api.resources.OperationOutcome;
-import gov.va.dvp.cdw.xsd.model.CdwEncounter101Root;
-import java.util.Collections;
-import java.util.function.Function;
-import javax.validation.constraints.Min;
-import lombok.AllArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.util.MultiValueMap;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-/**
- * Request Mappings for Encounter, see http://www.hl7.org/fhir/DSTU2/encounter.html for
- * implementation details.
- */
-@SuppressWarnings("WeakerAccess")
-@Validated
-@RestController
-@RequestMapping(
-  value = {"/dstu2/Encounter"},
-  produces = {"application/json", "application/json+fhir", "application/fhir+json"}
-)
-@AllArgsConstructor(onConstructor = @__({@Autowired}))
-public class EncounterController {
-  private Transformer transformer;
-  private MrAndersonClient mrAndersonClient;
-  private Dstu2Bundler bundler;
-
-  private Encounter.Bundle bundle(MultiValueMap<String, String> parameters, int page, int count) {
-    CdwEncounter101Root root = search(parameters);
-    LinkConfig linkConfig =
-        LinkConfig.builder()
-            .path("Encounter")
-            .queryParams(parameters)
-            .page(page)
-            .recordsPerPage(count)
-            .totalRecords(root.getRecordCount().intValue())
-            .build();
-    return bundler.bundle(
-        BundleContext.of(
-            linkConfig,
-            root.getEncounters() == null
-                ? Collections.emptyList()
-                : root.getEncounters().getEncounter(),
-            transformer,
-            Encounter.Entry::new,
-            Encounter.Bundle::new));
-  }
-
-  /** Read by id. */
-  @GetMapping(value = {"/{publicId}"})
-  public Encounter read(@PathVariable("publicId") String publicId) {
-    return transformer.apply(
-        firstPayloadItem(
-            hasPayload(search(Parameters.forIdentity(publicId)).getEncounters()).getEncounter()));
-  }
-
-  /**
-   * The XML should remain the same, but the version of the resource needs to be incremented for
-   * SQL52.
-   */
-  private CdwEncounter101Root search(MultiValueMap<String, String> params) {
-    Query<CdwEncounter101Root> query =
-        Query.forType(CdwEncounter101Root.class)
-            .profile(Query.Profile.DSTU2)
-            .resource("Encounter")
-            .version("1.02")
-            .parameters(params)
-            .build();
-    return hasPayload(mrAndersonClient.search(query));
-  }
-
-  /** Search by _id. */
-  @GetMapping(params = {"_id"})
-  public Encounter.Bundle searchById(
-      @RequestParam("_id") String id,
-      @RequestParam(value = "page", defaultValue = "1") @Min(1) int page,
-      @CountParameter @Min(0) int count) {
-    return bundle(
-        Parameters.builder().add("identifier", id).add("page", page).add("_count", count).build(),
-        page,
-        count);
-  }
-
-  /** Search by Identifier. */
-  @GetMapping(params = {"identifier"})
-  public Encounter.Bundle searchByIdentifier(
-      @RequestParam("identifier") String id,
-      @RequestParam(value = "page", defaultValue = "1") @Min(1) int page,
-      @CountParameter @Min(0) int count) {
-    return bundle(
-        Parameters.builder().add("identifier", id).add("page", page).add("_count", count).build(),
-        page,
-        count);
-  }
-
-  /** Hey, this is a validate endpoint. It validates. */
-  @PostMapping(
-    value = "/$validate",
-    consumes = {"application/json", "application/json+fhir", "application/fhir+json"}
-  )
-  public OperationOutcome validate(@RequestBody Encounter.Bundle bundle) {
-    return Dstu2Validator.create().validate(bundle);
-  }
-
-  public interface Transformer
-      extends Function<CdwEncounter101Root.CdwEncounters.CdwEncounter, Encounter> {}
-}
